feat: launch file newdance multi robot
